move images to image folder
